Move handling for rm command into separate file.
